Move lay_out_cols to view.
